#4 change wallet transaction amount value type to double

+++ src/main/kotlin/pl/szymonleyk/betgame/wallettransactions/WalletTransactionData.kt
@@ -2,4 +2,7 @@ package pl.szymonleyk.betgame.wallettransactions
 
 import java.time.LocalDate
 
-data class WalletTransactionData(val transactionDate: LocalDate, val amount: Int)
+data class WalletTransactionData(
+    val transactionDate: LocalDate,
+    val amount: Double
+)
